requirements.txt
django-allauth
django-allauth-ui
django-auto-prefetch
django-debug-toolbar
django-fastdev
django-htmx
django-machina
django-model-utils
django-money
django-oscar
django-phonenumber-field[phonenumbers]
django-redis
django-silk
django-taggit
django-template-partials
django-user-agents
django>=5.1
git+https://github.com/Naggafin/django-extensions.git
git+https://github.com/Naggafin/django-htmx-utils.git
git+https://github.com/mozilla/django-csp.git
git+https://www.github.com/Naggafin/django-view-breadcrumbs.git@improvements
hiredis
isort
jwt
nplusone
nplusone
parameterized
pillow
pinax-messages
pip
psycopg[c]
psycopg_pool
pycountry
python-dotenv
python-telegram-handler
redis
ruff
sorl-thumbnail
uv
wheel